+++ pydata.c
@@ -17,7 +17,7 @@
 #define EXTERN
 #include "pydata.h"
 
-smallint maxinbox[] = {0,0,0,0,1,2,2,2};
+int maxinbox[] = {0,0,0,0,1,2,2,2};
 unsigned int boxsize = sizeof maxinbox / sizeof maxinbox[0];
 
 killer_state const null_killer_state = {
